+++ python-bindings/b2_terraform/provider_tool.py
@@ -230,14 +230,18 @@ class Bucket(Command):
         default_retention = params.pop('default_retention', None)
         bucket = self.api.create_bucket(**params)
         if default_retention is not None:
-            params = self._preprocess(
-                file_lock_configuration=file_lock_configuration,
-                bucket_type=bucket_type,
-                bucket_info=bucket_info,
-            )
-            params.pop('is_file_lock_enabled', None)  # this can only be set during bucket creation
-            bucket = bucket.update(**params)
-
+            try:
+                params = self._preprocess(
+                    file_lock_configuration=file_lock_configuration,
+                    bucket_type=bucket_type,
+                    bucket_info=bucket_info,
+                )
+                params.pop('is_file_lock_enabled', None)  # this can only be set during bucket creation
+                bucket = bucket.update(**params)
+            except Exception:
+                self.api.delete_bucket(bucket)
+                raise
+        
         return self._postprocess(bucket)
 
     def resource_read(self, *, bucket_id, **kwargs):
@@ -291,6 +295,10 @@ class Bucket(Command):
                 cors_rules[index] = change_keys(item, converter=camelize)
 
         for file_lock_configuration in kwargs.pop('file_lock_configuration', ()):
+            lock_enabled = file_lock_configuration.get('is_file_lock_enabled')
+            default_retention_set = file_lock_configuration.get('default_retention') is not None
+            if default_retention_set and not lock_enabled:
+                raise RuntimeError('default_retention can only be set if is_file_lock_enabled is true')
             if 'is_file_lock_enabled' in file_lock_configuration:
                 kwargs['is_file_lock_enabled'] = file_lock_configuration['is_file_lock_enabled']
             for default_retention in file_lock_configuration.get('default_retention', ()):